ETL Design Example (#1)
**Description**
While working at MadeUp Co. we were faced with a legacy ETL tool that was
not designed to be used as an enterprise utility. Initially designed as
a one-off import tool the project was poorly maintained and ill-equipped
to handle multiple customers coming from multiple competitor database
architectures.

Rather than start from scratch, I audited the design line by line to
surface deficits and successes in the current code. I assessed its
faults and defined borders within the code that we could extract and
reuse or replace. To summarize my findings and get the team working in
the same direction I outlined the results and paths forward within an
architectural review/design document. This is an example of what that
looks like.

The initial design doc was then broken up into specific phases such that
each member of the team could work towards a deliverable with minimal
downtime.