Calendar Sync MVP (#2)
**Description**
While working for MadeUp Co. an initial implementation of a 3rd party
integration was in place that was convoluted and mixed much of the
business logic with the functional logic. It was originally implemented
with no requirements. In addition, it was not being used by any
customers, which left a lot of partially implemented code peppered
throughout the project.

I introduced a standardized HTTP REST library in a previous feature that
ensured consistent response objects across 3rd parties. Therefore, I
outlined a work brief to bring the calendar sync implementation up to
par with the new integrations. This required surgical removal of any
unused code and reconfiguring routes to be more RESTful.