Telephony through application
**Description**
The customers wanted to have the ability to auto-populate details to a
form based on the in-process phone call. In order to support this I
leveraged Twilio to provide telephony in a RESTful way. Ultimately,
this initial design was more than we anticipated and could commit to, so
I worked with our product team and the customer to better evaluate their
needs and present them with a solution that worked with far less
overhead. The crux of the issue was that in order to rout calls
directly to a logged in user we would have had to purchase and leverage
individual tokens and phone numbers that we could associate with a user.