Technical Research Document (#4)
**Description**
I don't like making decisions in a bubble, specially when it involves
company money. To better facilitate asynchronous decision making, I
typically like to draw up a R&D document that outline the trade offs of
our top options. This document is handed over to our product team,
discussions are had and ultimately a path forward is decided. This
ensure that all stakeholders have say in the outcome and there are no
surprises.

The result of this document will later be turned into a technical
implementation outline that can be used by any developer to get started
on the implementation.